+++ packages/contracts/contracts/BalanceManager.sol
@@ -112,6 +112,10 @@ contract BalanceManager is CommitmentTreeManager {
     /// @notice Gather reserved gas assets and pay bundler calculated amount.
     /// @dev Bundler can be paid less than reserved amount. Reserved amount is refunded to user's
     /// stealth address in this case.
+    /// @dev If the amount of gas asset remaining after bundler payout is less than the operation's
+    ///      gasAssetRefundThreshold, we just give the remaining amount to bundler. This is because
+    ///      the cost of handling refund for dust note and later proving ownership of the dust note
+    ///      will outweigh the actual value of the note.
     /// @param op Operation, which contains info on how much gas was reserved
     /// @param opResult OperationResult, which contains info on how much gas was actually spent
     /// @param perJoinSplitVerifyGas Gas cost of verifying a single joinSplit proof
@@ -123,16 +127,20 @@ contract BalanceManager is CommitmentTreeManager {
         address bundler
     ) internal {
         EncodedAsset calldata encodedGasAsset = op.encodedGasAsset;
-        uint256 gasAssetAmount = op.maxGasAssetCost(perJoinSplitVerifyGas);
+        uint256 gasAssetReserved = op.maxGasAssetCost(perJoinSplitVerifyGas);
 
-        if (gasAssetAmount > 0) {
-            // Request reserved gasAssetAmount from teller.
+        if (gasAssetReserved > 0) {
+            // Request reserved gasAssetReserved from teller.
             /// @dev This is safe because _processJoinSplitsReservingFee is
-            /// guaranteed to have reserved gasAssetAmount since it didn't throw.
-            _teller.requestAsset(encodedGasAsset, gasAssetAmount);
+            /// guaranteed to have reserved gasAssetReserved since it didn't throw.
+            _teller.requestAsset(encodedGasAsset, gasAssetReserved);
 
             uint256 bundlerPayout = OperationUtils
                 .calculateBundlerGasAssetPayout(op, opResult);
+            if (gasAssetReserved - bundlerPayout < op.gasAssetRefundThreshold) {
+                bundlerPayout = gasAssetReserved; // Give all to bundler if under threshold
+            }
+
             AssetUtils.transferAssetTo(encodedGasAsset, bundler, bundlerPayout);
         }
     }