+++ packages/contracts/CHANGELOG.md
@@ -2,7 +2,11 @@
 
 ### Unreleased
 
+- BalanceManager checks if reserved - payout < threshold, if so it just gives it all to bundler so there's no refund
+- add op.gasAssetRefundThreshold to operation
+- move `HandlerBase` into helpers so OperationGenerator can also use
 - fix bug in CTM require statement for checking batch is non empty
+- move `InvariantUtils` into helpers so OperationGenerator can also use
 - modify CTM/OffchainMerkle invariant handler to use seed then generate rand struct inputs rather than potentially huge array of structs which slows down test time
 - group processJoinSplit transfers by contiguous subarray to save 10k gas per joinsplit of same asset
 - replace `handleRefundNote` with `handleRefundNotes`, which inserts notes to tree's batch array in one go to save on SSTOREs